backend/requirements.txt
bcrypt==4.0.1
cryptography==41.0.7
fastapi==0.108.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
SQLAlchemy==2.0.25
uvicorn==0.25.0
